notifications/admin/requirements.txt
Django
python-dotenv
django-summernote
django-tinymce
django-stubs
psycopg2
uwsgi
gunicorn


